[refactor] convert services and utils to TS
